move test_008_hactl to test_14_failover_related
